+++ app/src/main/java/com/napzak/market/presentation/home/component/HomeRecommendationItemGroup.kt
@@ -28,8 +28,8 @@ import kotlinx.collections.immutable.toImmutableList
 @Composable
 fun HomeRecommendationItemGroup(
     recommendedItems: ImmutableList<ProductItem>,
-    onLikeClick: (Int) -> Unit,
-    onItemClick: (Int) -> Unit,
+    onLikeClick: (Long) -> Unit,
+    onItemClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     val context by rememberUpdatedState( LocalContext.current )
@@ -106,7 +106,7 @@ private fun HomeRecommendationGroupPreview() {
                 repeat(4) {
                     this.add(
                         ProductItem(
-                            productId = it,
+                            productId = it.toLong(),
                             productName = "딸기 마이멜로디 마스코트 인형",
                             genreName = "산리오",
                             price = 35000,
